Move Fdml class into its own file
